Add generic MongoDB CRUD service and wizard registration

+++ DiagonAlley2.0/Models/Wizard.cs
@@ -6,21 +6,23 @@ namespace DiagonAlley2._0.Models
 {
     public class Wizard
     {
-
         [BsonId]
         [BsonRepresentation(BsonType.ObjectId)]
         public string? Id { get; set; }
+
         public string Name { get; set; } = string.Empty;
         public string Password { get; set; } = string.Empty;
 
         [BsonRepresentation(BsonType.String)]
         public WizardLevel Level { get; set; } = WizardLevel.Regular;
-        public double Discount { get; set; } = 0.0;
+
+        public decimal Discount { get; set; } = 0.0m;
+
         public List<CartItem> Cart { get; set; } = new();
+
         public bool IsAdmin { get; set; } = false;
 
-
-        public double CartTotal()
+        public decimal CartTotal()
         {
             return Cart.Sum(c => c.TotalPrice);
         }
@@ -29,12 +31,11 @@ namespace DiagonAlley2._0.Models
         {
             Discount = Level switch
             {
-                WizardLevel.Gold => 0.30,
-                WizardLevel.Silver => 0.20,
-                WizardLevel.Bronze => 0.10,
-                _ => 0.0
+                WizardLevel.Gold => 0.30m,
+                WizardLevel.Silver => 0.20m,
+                WizardLevel.Bronze => 0.10m,
+                _ => 0.00m
             };
         }
-
     }
 }
